+++ application/model.py
@@ -10,12 +10,9 @@ session = db.session
 
 class User(db.Model):
     __tablename__ = "user"
-    __table_args__ = (
-        db.UniqueConstraint("email", name="unique_user")
-    )
 
     uuid = db.Column(UUIDType, primary_key=True, default=uuid.uuid4, nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.datetime.utcnow)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
 
     username = db.Column(db.String(80), unique=True, nullable=False)
     email = db.Column(EmailType(), unique=True, nullable=False)
