+++ application/wsgi.py
@@ -1,9 +1,6 @@
-from waitress import serve
-
-from application import create_app
-
+from application.app import create_app
 
 
 if __name__ == "__main__":
     application = create_app()
-    serve(application)
+    application.run()
