+++ docker-compose.yml
@@ -13,6 +13,5 @@ services:
     working_dir: /src/
     environment:
       PYTHONPATH: /src
-    env_file:
-      - credentials.env
+      DATABASE_URL: sqlite:///../db.sqlite3
     command: ["python", "application/wsgi.py"]
