+++ requirements.txt
@@ -34,5 +34,6 @@ sphinxcontrib-websupport==1.1.0
 SQLAlchemy==1.3.7
 SQLAlchemy-Utils==0.33.9
 urllib3==1.24.3
+uWSGI==2.0.18
 waitress==1.2.1
 Werkzeug==0.14.1
